pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-skills"
version = "0.1.0"
description = "Runtime for executing reusable AI agent skills and capability bindings"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [{name = "Guillermo Fernandez Alvarez", email = "gfernandf+as@gmail.com"}]
requires-python = ">=3.11"
keywords = ["ai", "agents", "skills", "capabilities", "workflows", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pyyaml>=6.0",
    "requests>=2.28",
]

[project.optional-dependencies]
pdf = ["PyPDF2>=3.0"]
web = ["beautifulsoup4>=4.12"]
otel = [
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-exporter-otlp>=1.20",
]
all = ["PyPDF2>=3.0", "beautifulsoup4>=4.12"]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.1",
    "pytest-benchmark>=4.0",
    "PyPDF2>=3.0",
    "beautifulsoup4>=4.12",
    "pip-audit>=2.6",
    "hypothesis>=6.80",
    "mutmut>=2.4",
    "ruff>=0.4",
]

[project.scripts]
agent-skills = "cli.main:main"

[project.entry-points."agent_skills.auth"]
# Pluggable auth backends — third parties register here
# api_key = "runtime.auth:ApiKeyStore"

[project.entry-points."agent_skills.invoker"]
# Protocol invokers (openapi, mcp, pythoncall, etc.)
# Custom protocols register their invoker classes here

[project.entry-points."agent_skills.binding_source"]
# Binding/service sources: official, community, local
# community = "my_plugin.bindings:CommunitySource"

[project.urls]
Homepage = "https://github.com/gfernandf/agent-skills"
Repository = "https://github.com/gfernandf/agent-skills"
Registry = "https://github.com/gfernandf/agent-skill-registry"

[tool.setuptools.packages.find]
include = [
    "cli*",
    "runtime*",
    "gateway*",
    "customer_facing*",
    "customization*",
    "official_services*",
    "official_mcp_servers*",
]

[tool.pytest.ini_options]
testpaths = ["runtime", "."]
addopts = [
    "--cov=runtime",
    "--cov=gateway",
    "--cov=customer_facing",
    "--cov=customization",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html:artifacts/htmlcov",
    "--cov-fail-under=75",
]

[tool.coverage.run]
branch = true
source = ["runtime", "gateway", "customer_facing", "customization"]
omit = [
    "*/test_*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.",
    "raise NotImplementedError",
    "@overload",
]
show_missing = true

[tool.ruff.lint]
ignore = ["E741"]

[tool.ruff.lint.per-file-ignores]
"cli/main.py" = ["E402"]
"test_*.py" = ["E402"]
"validate_bindings.py" = ["E402"]
"rename_capabilities.py" = ["E402"]
"fix_bindings.py" = ["E402"]
"run_pdf_test.py" = ["E402"]
"skills.py" = ["E402"]
"_run_tests.py" = ["E402"]
"_runner.py" = ["E402"]
"examples/**" = ["E402"]
"tooling/**" = ["E402"]
